Save the changes required for BPFCov support
1. Upgrade LLVM to 12

Somehow, current BPFCov only supports LLVM 12 (I'm not that much
familier with the LLVM, but seems they are relying on the feature of
"new" PassManager introduced in LLVM 12.

https://github.com/elastic/bpfcov/blob/main/CMakeLists.txt#L48

2. Add `opt` CLI tool

An official BPFCov instruction says that we need to execute their custom
LLVM pass with `opt` CLI tool. Build it together with Clang.

3. Build libBPFCov.so and expose it as an artifact

Need later in the cilium-runtime.

Signed-off-by: Yutaro Hayakawa <yutaro.hayakawa@isovalent.com>